[#16]: Update existings posts with option `--update`
Add ability to command line to specify the option
to overwrite the content. By default if the local article exists,
it would not be updated.

Example:

```shell
crystal run src/cli.cr -- -u miry -d posts/medium_miry --update
```

+++ src/medup/tool.cr
@@ -10,12 +10,13 @@ module Medup
 
     user : String?
 
-    def initialize(token : String?, @user : String?, dist : String?, format : String?, source : String?)
+    def initialize(token : String?, @user : String?, dist : String?, format : String?, source : String?, update : Bool?)
       @client = Medium::Client.new(token, @user)
       Medium::Client.default = @client
       @dist = (dist || DIST_PATH).as(String)
       @source = (source || SOURCE_AUTHOR_POSTS).as(String)
       @format = (format || MARKDOWN_FORMAT).as(String)
+      @update = update.nil? ? false : update.not_nil!
     end
 
     def backup
@@ -56,6 +57,7 @@ module Medup
       end
 
       if File.exists?(filepath)
+        return unless @update
         File.delete(filepath + ".old") if File.exists?(filepath + ".old")
         File.rename(filepath, filepath + ".old")
       end
